Made it passed first seg fault
I think it was an error having to do with pass by copy vs by reference in your allocateMemory function. Basically it got a copy of the struct allocated it, and then the copy dies. That is my guess at least. I got it to work when I switched it to a Matrix* and did some reformatting